Massive update
Refactored code so that everything is database neutral. As a result, the
schema was moved to migrations (using the 0.5.0 schema as the initial
schema) and should mean that the 3 new schema files work. Also created
fixtures to test.

Removed the config files. Only the templates should remain. This is to
ensure configurations are not committed in with passwords accidentaly.

Fixes #298, #269, #304, and hopefully #90 and #287.